Update start-phase-framework.md
v1.2

This update fixes the creation of files and folders in the project.

Added a function that allows the AI to make recommendations on project built-out based on the answers given in step 1. You can now accept the AI recommendation and have it build out the entire project structure or manually go through the build-out process.